put the initial template on a diet
