+++ {{ project_slug }}/.pre-commit-config.yaml
@@ -1,33 +1,34 @@
+---
+# See https://pre-commit.com for more information
+# See https://pre-commit.com/hooks.html for more hooks
 repos:
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.4.0
+    hooks:
+      - id: trailing-whitespace
+      - id: end-of-file-fixer
+      - id: check-yaml
+      - id: check-added-large-files
   - repo: https://github.com/charliermarsh/ruff-pre-commit
     # Ruff version.
-    rev: 'v0.0.245'
+    rev: 'v0.0.261'
     hooks:
       - id: ruff
-        args: [--fix-only, -e]
-
-  - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.1.0
+        args: [--fix, --exit-non-zero-on-fix]
+  - repo: https://github.com/adrienverge/yamllint
+    rev: v1.29.0
     hooks:
-      - id: check-merge-conflict
-      - id: check-toml
-      - id: check-yaml
-      - id: debug-statements
-      - id: end-of-file-fixer
-      - id: mixed-line-ending
-      - id: trailing-whitespace
-
-  - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v0.931
-    hooks:
-      - id: mypy
-        exclude: ^(tests/|.angreal/|conftest.py)
-        additional_dependencies: [types-aiofiles]
-
+      - id: yamllint
+        name: Check YAML files with yamllint
+        entry: yamllint -c .yamllint-config.yml --strict
+        types: [yaml]
   - repo: https://github.com/codespell-project/codespell
-    rev: v2.1.0
+    rev: v2.2.2
     hooks:
       - id: codespell
         name: Run codespell to check for common misspellings in files
         language: python
         types: [text]
+        exclude: ^docs/themes/.*$
+        args:
+          - --ignore-words=.spelling_wordlist.txt
