+++ {{ project_slug }}/.yamllint-config
@@ -1,0 +1,7 @@
+---
+# https://yamllint.readthedocs.io/en/stable/
+extends: default
+
+rules:
+  line-length:
+    max: 110
