--- {{ project_slug }}/docs/archetypes/default.md
@@ -1,6 +1,0 @@
----
-title: {% raw %}"{{ replace .Name "-" " " | title }}"{% endraw %}
-date: {% raw %}{{ .Date }}{% endraw %}
-draft: true
----
-
