--- {{ project_slug }}/docs/config.toml
@@ -1,76 +1,0 @@
-baseURL = "https://{{project_slug}}.github.io/{{ project_slug }}/"
-
-languageCode = "en-us"
-DefaultContentLanguage = "en"
-title = "{{ project_name }}"
-theme = "hugo-theme-learn"
-pygmentsCodeFences = true
-pygmentsStyle = "github-dark"
-
-defaultContentLanguage = "en"
-defaultContentLanguageInSubdir= false
-enableMissingTranslationPlaceholders = false
-
-
-
-
-
-
-[params]
-  editURL = "https://github.dev/{{ github_username }}/{{ project_slug }}/blob/main/docs/"
-  # Author of the site, will be used in meta information
-  author = "{{ author_name }}"
-  # Description of the site, will be used in meta information
-  description = "Documentation for the Angreal software package."
-  # Shows a checkmark for visited pages on the menu
-  showVisitedLinks = false
-  # Disable search function. It will hide search bar
-  disableSearch = false
-  # Javascript and CSS cache are automatically busted when new version of site is generated.
-  # Set this to true to disable this behavior (some proxies don't handle well this optimization)
-  disableAssetsBusting = false
-  # Set this to true to disable copy-to-clipboard button for inline code.
-  disableInlineCopyToClipBoard = false
-  # A title for shortcuts in menu is set by default. Set this to true to disable it.
-  disableShortcutsTitle = false
-  # If set to false, a Home button will appear below the search bar on the menu.
-  # It is redirecting to the landing page of the current language if specified. (Default is "/")
-  disableLandingPageButton = true
-  # When using mulitlingual website, disable the switch language button.
-  disableLanguageSwitchingButton = true
-  # Hide breadcrumbs in the header and only show the current page title
-  disableBreadcrumb = true
-  # If set to true, prevents Hugo from including the mermaid module if not needed (will reduce load times and traffic)
-  disableMermaid = false
-  # Specifies the remote location of the mermaid js
-  customMermaidURL = "https://unpkg.com/mermaid@8.8.0/dist/mermaid.min.js"
-  # Hide Next and Previous page buttons normally displayed full height beside content
-  disableNextPrev = true
-  # Order sections in menu by "weight" or "title". Default to "weight"
-  ordersectionsby = "weight"
-  # Change default color scheme with a variant one. Can be "red", "blue", "green".
-  themeVariant = "blue"
-  # Provide a list of custom css files to load relative from the `static/` folder in the site root.
-  #custom_css = ["css/foo.css", "css/bar.css"]
-  # Change the title separator. Default to "::".
-  #titleSeparator = "-"
-
-
-
-[outputs]
-home = [ "HTML", "RSS", "JSON"]
-
-
-
-[[menu.shortcuts]]
-name = "<i class='fab fa-gitlab'></i> Git Repo"
-identifier = "ds"
-url = "https://github.com/{{ github_username }}/{{ project_slug }}"
-weight = 10
-
-
-[[menu.shortcuts]]
-name = "<i class='fa fa-download'></i> Download"
-url = "https://github.com/{{ github_username }}/{{ project_slug }}/archive/refs/heads/main.zip"
-weight = 20
-
