--- {{ project_slug }}/docs/content/_index.md
@@ -1,10 +1,0 @@
----
-title: {{ project_name }} Home
-type: page
----
-
-# {{ project_name }}
---- 
-
-
-The next big thing !
