--- {{ project_slug }}/docs/layouts/partials/logo.html
@@ -1,3 +1,0 @@
-<h1> {{ package_name }} </h1>
-<div></div>
-<div></div>
