+++ {{ project_slug }}/setup.cfg
@@ -7,7 +7,7 @@ author_email = {{ author_email }}
 description = {{ project_short_description }}
 long_description = file: README.md
 long_description_content_type = text/markdown
-license = "Apache License 2.0"
+license = ""
 classifiers = 
   "Development Status :: 4 - Beta",
   "Intended Audience :: Developers",
@@ -21,9 +21,9 @@ classifiers =
   "Topic :: Software Development :: Libraries :: Python Modules",
 
 project_urls =
-    Homepage="https://{{ github_username }}.github.io/{{ project_slug }}"
-    Source Code="https://{{ github_username }}.github.io/{{ project_slug }}"
-    Documentation="https://{{ github_username }}.github.io/{{ project_slug }}"
+    Homepage=""
+    Source Code=""
+    Documentation=""
 
 [options]
 python_requires = ~=3.7
