+++ {{ project_slug }}/{{ package_name }}/__init__.py
@@ -1,8 +1,4 @@
-"""
-    {{ package_name }}
-    ~~~~~~~~~~~~~~~~~~~~~~~
-
-     {{ package_name }}
+"""{{ package_name }}
 
 """
 __version__ = "0.0.0a"
