+++ {{ project_slug }}/{{ package_name }}/__main__.py
@@ -1,11 +1,6 @@
-"""
-    {{ package_name }}.__main__
-    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-    {{ package_name }}'s registered entry point
+"""{{ package_name }}'s registered entry point.
 
 """
-
 
 def main() -> None:
     pass
